% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_midas2s.R
\name{platform_midas2s}
\alias{platform_midas2s}
\title{An Bayesian platform design without subgroup efficacy exploration(midas-2s), which is the degenerate competing design in the simulation.}
\usage{
platform_midas2s(seed, p, p_tox, C_T = 0.85, C_E1 = 0.15, C_E2 = 0.999)
}
\arguments{
\item{seed}{set a random seed to maintain the repeatability of the simulation results.}

\item{p}{a matrix indicating the efficacy. Row number represents the number of candidate drugs.}

\item{p_tox}{a vector indicating the toxicity.}

\item{C_T}{early toxicity stopping threshold, which refers to a predefined threshold used to determine when a clinical trial should be stopped early due to unacceptable levels of toxicity or adverse events in the study participants. This threshold is established to ensure the safety and well-being of the trial participants and to prevent further harm.}

\item{C_E1}{early futility stopping threshold, which refers to a predefined threshold used to determine when a clinical trial should be stopped early due to lack of efficacy or futility. It is established to prevent the continuation of a trial that is unlikely to demonstrate a significant treatment effect, thus saving time, resources, and participant exposure to ineffective treatments.}

\item{C_E2}{early efficacy stopping threshold, which refers to a predefined threshold used to determine when a clinical trial should be stopped early due to the demonstration of significant efficacy or positive treatment effects. This threshold is established to allow for timely decision-making and saves sample size.}
}
\value{
term.tox the indicator of whether early stopping for toxicity

term.fut the indicator of whether early stopping for futility

term.eff the indicator of whether early stopping for efficacy

final.eff a vector of final decision, either efficacy or inefficacy

N sample size, which refers to the number of participants included in a study or experiment.
}
\description{
MIDAS-2s is the degenerate competing designs that do not consider subgroups. Beta-binomial model is applied for efficacy in whole population of each arm.
}
\examples{


  # Example 1
  p0 <- c(0.1,0.1,0.1,0.1)
  p1 <- c(0.1,0.1,0.1,0.1)

  p <- rbind(p0,p1)
  p_tox <- c(0.1,0.4)

  # consider 1 candidate drugs with 4 subgroups
  result <- platform_midas2s(seed=20,p,p_tox,C_T=0.85,C_E1=0.15,C_E2=0.999)
  result


  \donttest{
  # Example 2
  p0 <- c(0.05,0.10,0.05,0.10)
  p1 <- c(0.24,0.40,0.12,0.22)
  p2 <- c(0.24,0.40,0.12,0.22)
  p3 <- c(0.12,0.22,0.05,0.10)
  p4 <- c(0.24,0.40,0.12,0.22)
  p5 <- c(0.28,0.45,0.12,0.22)
  p6 <- c(0.24,0.40,0.12,0.22)
  p7 <- c(0.12,0.22,0.05,0.10)

  p <- rbind(p0, p1, p2, p3, p4, p5, p6, p7)
  p_tox <- c(0.10,0.10,0.10,0.10,0.10,0.10,0.15,0.20)

  # consider 7 candidate drugs with 4 subgroups
  result <- platform_midas2s(seed=12,p,p_tox,C_T=0.85,C_E1=0.15,C_E2=0.999)
  result
  }


}
