% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform_design_with_hierarchical_constrain.r
\name{hc_platform}
\alias{hc_platform}
\title{An Information Borrowing Drug-Combination Bayesian Platform Design(MIDAS-2)}
\usage{
hc_platform(seed, p, p_tox)
}
\arguments{
\item{seed}{set a random seed to maintain the repeatability of the simulation results.}

\item{p}{a matrix indicating the efficacy. Row number represents the number of candidate drugs.}

\item{p_tox}{a vector indicating the toxicity.}
}
\value{
term.tox the indicator of whether early stopping for toxicity

term.fut the indicator of whether early stopping for futility

term.eff the indicator of whether early stopping for efficacy

final.eff a vector of final decision, either efficacy or inefficacy

post.subg subgroup analysis for treatments

post.sign signature analysis for treatments

post.spike posterior estimation for spike parameters

best selection of best treatment for each subgroup
}
\description{
An Information borrowing drug-combination Bayesian platform design with subgroup exploration and hierarchical constrain.
}
\examples{
p0 <- c(    0.1,    0.1,    0.1,   0.1)
p1 <- c(    0.1,    0.1,    0.1,   0.1)
p2 <- c(    0.1,    0.1,    0.1,   0.1)
p3 <- c(    0.1,    0.1,    0.1,   0.1)
p4 <- c(    0.1,    0.1,    0.1,   0.1)
p5 <- c(    0.1,    0.1,    0.1,   0.1)
p6 <- c(    0.1,    0.1,    0.1,   0.1)
p7 <- c(    0.1,    0.1,    0.1,   0.1)
p <- rbind(p0, p1, p2, p3, p4, p5, p6, p7)
p_tox <- c(0.1,0.1,0.1,0.1,0.1,0.1,0.1,0.1)

\donttest{
# consider 7 candidate drugs with 4 subgroups
result <- hc_platform(seed=12,p,p_tox)
result
}

}
